Make EQ work on 64 bit sims again.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@9214 a1c6a512-1295-4272-9138-f99709370657

+++ apps/eq.c
@@ -102,7 +102,7 @@ static const unsigned long atan_table[] = {
  *         representing -1 and 1 respectively. 
  */
 long fsincos(unsigned long phase, long *cos) {
-    long x, x1, y, y1;
+    int32_t x, x1, y, y1;
     unsigned long z, z1;
     int i;
